Reorganizing input converters into helpers and adding validation logic
