+++ src/components/Launch/LaunchWorkflowForm/inputHelpers/constants.ts
@@ -1,0 +1,5 @@
+import { Core } from 'flyteidl';
+
+export function literalNone(): Core.ILiteral {
+    return { scalar: { noneType: {} } };
+}
