+++ src/components/Launch/LaunchWorkflowForm/inputHelpers/float.ts
@@ -1,0 +1,17 @@
+import { Core } from 'flyteidl';
+import { ConverterInput, InputHelper } from './types';
+
+function toLiteral({ value }: ConverterInput): Core.ILiteral {
+    const floatValue =
+        typeof value === 'number' ? value : parseFloat(value.toString());
+    return {
+        scalar: { primitive: { floatValue } }
+    };
+}
+
+function validate({ value }: ConverterInput) {}
+
+export const floatHelper: InputHelper = {
+    toLiteral,
+    validate
+};
