+++ src/components/Launch/LaunchWorkflowForm/inputHelpers/none.ts
@@ -1,0 +1,7 @@
+import { literalNone } from './constants';
+import { InputHelper } from './types';
+
+export const noneHelper: InputHelper = {
+    toLiteral: literalNone,
+    validate: () => {}
+};
